backend/requirements-pi.txt
flask==3.0.0
pillow==10.0.0
qrcode==7.4.2
python-dotenv==1.0.0
opencv-python==4.8.0.76
numpy==1.24.3
picamera2==0.3.12
pyzbar==0.1.9
paho-mqtt==1.6.1 